Add files via upload
K-Nearest Neighbors (KNN) is a supervised learning algorithm used for classification and regression. It is a simple, non-parametric, and instance-based algorithm that makes predictions by finding the K most similar data points (neighbors) to a given input.

exampl:-
      Imagine you're in a new school 🏫 and you don’t know anyone. You want to figure out which friend group you should join.

You look around and notice students wearing similar clothes, speaking in a certain way, or having common interests.
You decide to join the group that has the most people who are similar to you.
This is exactly how KNN works! It finds the K most similar data points and assigns the new point to the most common category.